+++ frontend/openchat-agent/src/services/user/candid/types.d.ts
@@ -1316,7 +1316,7 @@ export type RemoveReactionResponse = { 'MessageNotFound' : null } |
   { 'UserSuspended' : null } |
   { 'SuccessV2' : PushEventResult };
 export interface ReplyContext {
-  'chat_if_other' : [] | [[MultiUserChat, [] | [MessageIndex]]],
+  'chat_if_other' : [] | [[Chat, [] | [MessageIndex]]],
   'event_list_if_other' : [] | [[ChatId, [] | [MessageIndex]]],
   'event_index' : EventIndex,
 }